requirements.txt
chromadb
flask
langchain-classic
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
pandas
python-dotenv
sentence-transformers
requests
